+++ homeassistant/components/automation/__init__.py
@@ -43,40 +43,35 @@ from homeassistant.helpers.script import (
     ATTR_MODE,
     CONF_MAX,
     CONF_MAX_EXCEEDED,
-    SCRIPT_MODE_SINGLE,
     Script,
-    make_script_schema,
 )
 from homeassistant.helpers.script_variables import ScriptVariables
 from homeassistant.helpers.service import async_register_admin_service
-from homeassistant.helpers.singleton import singleton
 from homeassistant.helpers.trigger import async_initialize_triggers
 from homeassistant.helpers.typing import TemplateVarsType
 from homeassistant.loader import bind_hass
 from homeassistant.util.dt import parse_datetime
 
+from .config import async_validate_config_item
+from .const import (
+    CONF_ACTION,
+    CONF_CONDITION,
+    CONF_INITIAL_STATE,
+    CONF_TRIGGER,
+    DEFAULT_INITIAL_STATE,
+    DOMAIN,
+    LOGGER,
+)
+from .helpers import async_get_blueprints
+
 # mypy: allow-untyped-calls, allow-untyped-defs
 # mypy: no-check-untyped-defs, no-warn-return-any
 
-DOMAIN = "automation"
 ENTITY_ID_FORMAT = DOMAIN + ".{}"
 
-DATA_BLUEPRINTS = "automation_blueprints"
 
-CONF_DESCRIPTION = "description"
-CONF_HIDE_ENTITY = "hide_entity"
-
-CONF_CONDITION = "condition"
-CONF_ACTION = "action"
-CONF_TRIGGER = "trigger"
-CONF_CONDITION_TYPE = "condition_type"
-CONF_INITIAL_STATE = "initial_state"
 CONF_SKIP_CONDITION = "skip_condition"
 CONF_STOP_ACTIONS = "stop_actions"
-CONF_BLUEPRINT = "blueprint"
-CONF_INPUT = "input"
-
-DEFAULT_INITIAL_STATE = True
 DEFAULT_STOP_ACTIONS = True
 
 EVENT_AUTOMATION_RELOADED = "automation_reloaded"
@@ -87,37 +82,7 @@ ATTR_SOURCE = "source"
 ATTR_VARIABLES = "variables"
 SERVICE_TRIGGER = "trigger"
 
-_LOGGER = logging.getLogger(__name__)
-
 AutomationActionType = Callable[[HomeAssistant, TemplateVarsType], Awaitable[None]]
-
-_CONDITION_SCHEMA = vol.All(cv.ensure_list, [cv.CONDITION_SCHEMA])
-
-PLATFORM_SCHEMA = vol.All(
-    cv.deprecated(CONF_HIDE_ENTITY, invalidation_version="0.110"),
-    make_script_schema(
-        {
-            # str on purpose
-            CONF_ID: str,
-            CONF_ALIAS: cv.string,
-            vol.Optional(CONF_DESCRIPTION): cv.string,
-            vol.Optional(CONF_INITIAL_STATE): cv.boolean,
-            vol.Optional(CONF_HIDE_ENTITY): cv.boolean,
-            vol.Required(CONF_TRIGGER): cv.TRIGGER_SCHEMA,
-            vol.Optional(CONF_CONDITION): _CONDITION_SCHEMA,
-            vol.Optional(CONF_VARIABLES): cv.SCRIPT_VARIABLES_SCHEMA,
-            vol.Required(CONF_ACTION): cv.SCRIPT_SCHEMA,
-        },
-        SCRIPT_MODE_SINGLE,
-    ),
-)
-
-
-@singleton(DATA_BLUEPRINTS)
-@callback
-def async_get_blueprints(hass: HomeAssistant) -> blueprint.DomainBlueprints:  # type: ignore
-    """Get automation blueprints."""
-    return blueprint.DomainBlueprints(hass, DOMAIN, _LOGGER)  # type: ignore
 
 
 @bind_hass
@@ -194,7 +159,7 @@ def devices_in_automation(hass: HomeAssistant, entity_id: str) -> List[str]:
 
 async def async_setup(hass, config):
     """Set up the automation."""
-    hass.data[DOMAIN] = component = EntityComponent(_LOGGER, DOMAIN, hass)
+    hass.data[DOMAIN] = component = EntityComponent(LOGGER, DOMAIN, hass)
 
     await _async_process_config(hass, config, component)
 
@@ -263,7 +228,7 @@ class AutomationEntity(ToggleEntity, RestoreEntity):
         self._is_enabled = False
         self._referenced_entities: Optional[Set[str]] = None
         self._referenced_devices: Optional[Set[str]] = None
-        self._logger = _LOGGER
+        self._logger = LOGGER
         self._variables: ScriptVariables = variables
 
     @property
@@ -536,10 +501,12 @@ async def _async_process_config(
                 try:
                     config_block = cast(
                         Dict[str, Any],
-                        PLATFORM_SCHEMA(blueprint_inputs.async_substitute()),
+                        await async_validate_config_item(
+                            hass, blueprint_inputs.async_substitute()
+                        ),
                     )
                 except vol.Invalid as err:
-                    _LOGGER.error(
+                    LOGGER.error(
                         "Blueprint %s generated invalid automation with inputs %s: %s",
                         blueprint_inputs.blueprint.name,
                         blueprint_inputs.inputs,
@@ -561,7 +528,7 @@ async def _async_process_config(
                 script_mode=config_block[CONF_MODE],
                 max_runs=config_block[CONF_MAX],
                 max_exceeded=config_block[CONF_MAX_EXCEEDED],
-                logger=_LOGGER,
+                logger=LOGGER,
                 # We don't pass variables here
                 # Automation will already render them to use them in the condition
                 # and so will pass them on to the script.
@@ -600,7 +567,7 @@ async def _async_process_if(hass, config, p_config):
         try:
             checks.append(await condition.async_from_config(hass, if_config, False))
         except HomeAssistantError as ex:
-            _LOGGER.warning("Invalid condition: %s", ex)
+            LOGGER.warning("Invalid condition: %s", ex)
             return None
 
     def if_action(variables=None):
