+++ homeassistant/components/automation/helpers.py
@@ -1,0 +1,15 @@
+"""Helpers for automation integration."""
+from homeassistant.components import blueprint
+from homeassistant.core import HomeAssistant, callback
+from homeassistant.helpers.singleton import singleton
+
+from .const import DOMAIN, LOGGER
+
+DATA_BLUEPRINTS = "automation_blueprints"
+
+
+@singleton(DATA_BLUEPRINTS)
+@callback
+def async_get_blueprints(hass: HomeAssistant) -> blueprint.DomainBlueprints:  # type: ignore
+    """Get automation blueprints."""
+    return blueprint.DomainBlueprints(hass, DOMAIN, LOGGER)  # type: ignore
